Lift by-name Tool APIs to new type
